chore: update endpoints
refactor: structured codebase
